+++ lib/lite/uxid/reversible/scatterid.rb
@@ -6,12 +6,18 @@ module Lite
       class Scatterid < Base
 
         def encode
-          "#{coder_prefix}#{hash}"
+          swap
+          scatter
+
+          "#{coder_prefix}#{joined_array}"
         end
 
         def decode
           @id = id.delete_prefix(coder_prefix.to_s)
-          reverse_hash.to_i
+
+          unscatter
+          unswap
+          joined_array.to_i
         end
 
         private
@@ -20,26 +26,12 @@ module Lite
           @zero_padded_id ||= id.to_s.rjust(10, "0")
         end
 
-        # def coder_array
-        #   @coder_array ||= zero_padded_id.chars.collect(&:to_i)
-        # end
-
-        ###
-
-        def hash
-          swap
-          scatter
-          completed_string
+        def coder_array
+          @coder_array ||= zero_padded_id.chars.collect(&:to_i)
         end
 
-        def reverse_hash
-          unscatter
-          unswap
-          completed_string
-        end
-
-        def completed_string
-          @working_array.join
+        def joined_array
+          coder_array.join
         end
 
         def swapper_map(index)
@@ -48,34 +40,27 @@ module Lite
         end
 
         def swap
-          @working_array = zero_padded_id.chars.collect(&:to_i)
-          @working_array = @working_array.map.with_index do |digit, index|
-            swapper_map(index)[digit]
-          end
+          coder_array.map!.with_index { |digit, i| swapper_map(i)[digit] }
         end
 
-        # Reverse swap
         def unswap
-          @working_array = @working_array.collect.with_index do |digit, index|
-            swapper_map(index).rindex(digit)
-          end
+          coder_array.map!.with_index { |digit, i| swapper_map(i).rindex(digit) }
         end
 
         def scatter
-          sum_of_digits = @working_array.sum.to_i
-          @working_array = Array.new(10) do
-            @working_array.rotate!(coder_spin ^ sum_of_digits).pop
-          end
+          sum_of_digits = coder_array.sum.to_i
+          @coder_array = Array.new(10) { coder_array.rotate!(coder_spin ^ sum_of_digits).pop }
         end
 
         def unscatter
-          scattered_array = zero_padded_id.chars.collect(&:to_i)
-          sum_of_digits = scattered_array.inject(:+).to_i
-          @working_array = []
-          @working_array.tap do |unscatter|
+          scattered_array = coder_array
+          sum_of_digits = scattered_array.sum.to_i
+          @coder_array = []
+
+          coder_array.tap do |unscatter|
             10.times do
-              unscatter.push scattered_array.pop
-              unscatter.rotate! (sum_of_digits ^ coder_spin) * -1
+              unscatter.push(scattered_array.pop)
+              unscatter.rotate!((sum_of_digits ^ coder_spin) * -1)
             end
           end
         end