+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.0] - 2024-09-23
+### Added
+- Scatterid reversible lib
 ### Changed
 - Reorganize generators into better namespaces
 - Fixed issue where nanoid would return without the prefix